+++ src/hooks/api/useHoldingsNotes.js
@@ -1,8 +1,10 @@
-import { useNamespace, useOkapiKy } from '@folio/stripes/core';
 import { useQuery } from 'react-query';
 import { useIntl } from 'react-intl';
 import { useMemo } from 'react';
-import { OPTIONS, PARAMETERS_KEYS, MOD_INVENTORY_STORAGE } from '../../constants';
+
+import { useNamespace, useOkapiKy } from '@folio/stripes/core';
+
+import { OPTIONS, PARAMETERS_KEYS } from '../../constants';
 import { getMappedAndSortedNotes } from '../../utils/helpers';
 import { useErrorMessages } from '../useErrorMessages';
 
@@ -13,15 +15,16 @@ export const useHoldingsNotes = (options = {}) => {
   const ky = useOkapiKy();
   const [namespaceKey] = useNamespace({ key: HOLDINGS_NOTES_KEY });
   const { formatMessage } = useIntl();
-  const { showExternalModuleError } = useErrorMessages();
+  const path = 'holdings-note-types';
+  const { showExternalModuleError } = useErrorMessages({ path });
 
   const { data, isLoading: isHoldingsNotesLoading } = useQuery(
     {
       queryKey: [namespaceKey],
       cacheTime: Infinity,
       staleTime: Infinity,
-      queryFn: () => ky.get('holdings-note-types', { searchParams: { limit: 1000 } }).json(),
-      onError: (error) => showExternalModuleError(MOD_INVENTORY_STORAGE, error),
+      queryFn: () => ky.get(path, { searchParams: { limit: 1000 } }).json(),
+      onError: showExternalModuleError,
       ...options,
     },
   );