+++ src/hooks/api/useItemNotes.js
@@ -1,10 +1,13 @@
-import { useNamespace, useOkapiKy } from '@folio/stripes/core';
 import { useQuery } from 'react-query';
 import { useIntl } from 'react-intl';
 import { useMemo } from 'react';
-import { MOD_INVENTORY_STORAGE, OPTIONS, PARAMETERS_KEYS } from '../../constants';
+
+import { useNamespace, useOkapiKy } from '@folio/stripes/core';
+
+import { OPTIONS, PARAMETERS_KEYS } from '../../constants';
 import { getMappedAndSortedNotes } from '../../utils/helpers';
 import { useErrorMessages } from '../useErrorMessages';
+
 
 export const ITEM_NOTES_KEY = 'ITEM_NOTES_KEY';
 
@@ -12,15 +15,16 @@ export const useItemNotes = (options = {}) => {
   const ky = useOkapiKy();
   const [namespaceKey] = useNamespace({ key: ITEM_NOTES_KEY });
   const { formatMessage } = useIntl();
-  const { showExternalModuleError } = useErrorMessages();
+  const path = 'item-note-types';
+  const { showExternalModuleError } = useErrorMessages({ path });
 
   const { data, isLoading: isItemNotesLoading } = useQuery(
     {
       queryKey: [namespaceKey],
       cacheTime: Infinity,
       staleTime: Infinity,
-      queryFn: () => ky.get('item-note-types', { searchParams: { limit: 1000 } }).json(),
-      onError: (error) => showExternalModuleError(MOD_INVENTORY_STORAGE, error),
+      queryFn: () => ky.get(path, { searchParams: { limit: 1000 } }).json(),
+      onError: showExternalModuleError,
       ...options,
     },
   );