+++ src/hooks/api/usePatronGroup.js
@@ -3,22 +3,23 @@ import {
 } from 'react-query';
 
 import { useNamespace, useOkapiKy } from '@folio/stripes/core';
+
 import { useErrorMessages } from '../useErrorMessages';
-import { MOD_USERS } from '../../constants';
+
 
 export const PATRON_GROUP_KEY = 'PATRON_GROUP_KEY';
 
 export const usePatronGroup = (options = {}) => {
   const ky = useOkapiKy();
   const [namespaceKey] = useNamespace({ key: PATRON_GROUP_KEY });
-  const { showExternalModuleError } = useErrorMessages();
+  const { showExternalModuleError } = useErrorMessages({ path: 'groups' });
 
   const { data, isLoading } = useQuery(
     {
       queryKey: [namespaceKey],
       cacheTime: Infinity,
       staleTime: Infinity,
-      onError: (error) => showExternalModuleError(MOD_USERS, error),
+      onError: showExternalModuleError,
       queryFn: async () => {
         const { usergroups } = await ky.get('groups', { searchParams: { limit: 200 } }).json();