+++ src/components/BulkEditPane/BulkEditListSidebar/BulkEditListSidebar.js
@@ -1,5 +1,6 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
+import { useQueryClient } from 'react-query';
 
 import { ButtonGroup } from '@folio/stripes/components';
 import { buildSearch } from '@folio/stripes-acq-components';
@@ -11,12 +12,19 @@ import { IdentifierTab } from './IdentifierTab/IdentifierTab';
 import { QueryTab } from './QueryTab/QueryTab';
 import { LogsTab } from './LogsTab/LogsTab';
 import { useSearchParams } from '../../../hooks/useSearchParams';
+import { BULK_OPERATION_DETAILS_KEY } from '../../../hooks/api';
+import { RootContext } from '../../../context/RootContext';
 
 export const BulkEditListSidebar = () => {
   const history = useHistory();
   const location = useLocation();
   const { hasLogViewPerms, hasQueryPerms } = useBulkPermissions();
   const { criteria } = useSearchParams();
+  const queryClient = useQueryClient();
+  const {
+    setIsFileUploaded,
+    setVisibleColumns,
+  } = useContext(RootContext);
 
   const isQuery = criteria === CRITERIA.QUERY;
   const isLogs = criteria === CRITERIA.LOGS;
@@ -26,6 +34,12 @@ export const BulkEditListSidebar = () => {
     history.replace({
       search: buildSearch({ criteria: value }, location.search),
     });
+  };
+
+  const handleClearState = () => {
+    setIsFileUploaded(false);
+    setVisibleColumns(null);
+    queryClient.removeQueries({ queryKey: [BULK_OPERATION_DETAILS_KEY] });
   };
 
   return (
@@ -41,11 +55,11 @@ export const BulkEditListSidebar = () => {
 
       {/* IDENTIFIER TAB */}
       {isIdentifier && (
-        <IdentifierTab />
+        <IdentifierTab onClearState={handleClearState} />
       )}
 
       {/* QUERY TAB */}
-      {isQuery && hasQueryPerms && <QueryTab />}
+      {isQuery && hasQueryPerms && <QueryTab onClearState={handleClearState} />}
 
       {/* LOGS TAB */}
       {isLogs && <LogsTab />}