+++ src/components/BulkEditPane/BulkEditListSidebar/IdentifierTab/IdentifierTab.js
@@ -26,7 +26,7 @@ import { getIsDisabledByPerm } from '../utils/getIsDisabledByPerm';
 import { RootContext } from '../../../../context/RootContext';
 import { useErrorMessages } from '../../../../hooks/useErrorMessages';
 
-export const IdentifierTab = () => {
+export const IdentifierTab = ({ onClearState }) => {
   const history = useHistory();
   const location = useLocation();
   const permissions = useBulkPermissions();
@@ -35,7 +35,6 @@ export const IdentifierTab = () => {
   const {
     isFileUploaded,
     setIsFileUploaded,
-    setVisibleColumns,
   } = useContext(RootContext);
 
   const {
@@ -99,12 +98,10 @@ export const IdentifierTab = () => {
       }, history.location.search),
     });
 
-    setIsFileUploaded(false);
-    setVisibleColumns(null);
+    onClearState();
   }, [
     history,
-    setIsFileUploaded,
-    setVisibleColumns,
+    onClearState,
   ]);
 
   const handleCapabilityChange = (e) => {
@@ -120,8 +117,7 @@ export const IdentifierTab = () => {
       }, history.location.search),
     });
 
-    setVisibleColumns(null);
-    setIsFileUploaded(false);
+    onClearState();
   };
 
   const handleDragEnter = () => {