+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [UIBULKED-642](https://folio-org.atlassian.net/browse/UIBULKED-642) Rendering Instance record Electronic access in UI.
 * [UIBULKED-665](https://folio-org.atlassian.net/browse/UIBULKED-665) Show error link for central tenant when using bulk-edit query.
 * [UIBULKED-644](https://folio-org.atlassian.net/browse/UIBULKED-644) Rendering Instance record Subject data in UI.
+* [UIBULKED-649](https://folio-org.atlassian.net/browse/UIBULKED-649) Correct plurals on records matched and confirmation forms.
 
 ## [5.0.0](https://github.com/folio-org/ui-bulk-edit/tree/v5.0.0) (2025-03-12)